I'm absolutely done with this.
God damn it, this just made me remember how much I hate doing unnecessary effort for no payoff. I am done with this script. It kinda works and it kinda doesn't.

Currently known issues:

*Avatars won't de-round.
*After browsing repos, the style will reset until the next page load.
*Old style will be visible for a split second after loading the page.

If you want to fix any of that - go right ahead. But I'm not maintaining this or fixing this myself (for now at least), because I hated every second of writing this, and I just had enough. 

I miss those 15 minutes of my life that I spent on writing this...