New script: Bad Wineserver Killer + Wine Launch
This actually started out as me having to write custom launch parameters to a windows app I constantly use. And thus, I wrote a simple script with exports I need for the launch.

After that, I just launched the application via a bash script on my desktop. However, one day, this actually drove me completely insane as the amount of crashes it did on a hourly basis (which are not related to Wine, this crashes on windows a lot too) was such a headache and a chore, that constantly switching to desktop just to launch an application became a chore-y task in itself. Because of that, I have symlinked this script to my home folder, and I defined a shortcut so I could launch this app via a singular key on my keyboard. All good, right?... Well...

This piece of proprietary garbage™ actually started hanging wineserver in the process list, which made me unable to launch another wineserver until that other wineserver died. And so I added a special wineserver killer into the script, in case that limp-wristed twaddle of an application starts hoarding the wineserver again. And yes, this can cause data loss. I don't really care, because even without this, the application tends to have data losses periodically either way anyway. Use at your own risk.

Oh, and I hope somebody makes an open-source, or at very least Linux compatible alternative of it. Because being stuck with this proprietary garbage™ is making me lose not only my brain cells, but nerves too.