Update tsremover.user.js
I couldn't for the life of me figure out how to fire an event during url change, so I did with a dirty workaround - calling url clean function every 1 second. From what I've noticed, that did not degrade performance at all (on my end), even with like 30 and 40 youtube tabs open at the same time. 

Included run-at attribute to make the script fire as document starts.

Moved link cleaning into a function if I need to call it again.

Added check for "feature" in the url, since "feature" usually appears when you clicked on a user link (for example a video with timestamp for something), and therefore I don't see the need to clean the link. This might trigger anyway if "feature" goes away from the url and the function is ran, but I'll solve that at some point later.